moves update password to authentication folder
